chore(cli): Create fs utility module
